// frontend/src/api/generated/model/monitorResponseType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CloudOps API
 * Infrastructure resource management, monitoring and asynchronous tasks. Timestamps use UTC ISO-8601. Error responses share ApiError; errors contains field violations.
 * OpenAPI spec version: 0.0.1-SNAPSHOT
 */

export type MonitorResponseType =
  (typeof MonitorResponseType)[keyof typeof MonitorResponseType];

export const MonitorResponseType = {
  HTTP_CHECK: "HTTP_CHECK",
  PORT_CHECK: "PORT_CHECK",
  DNS_CHECK: "DNS_CHECK",
  PING: "PING",
  TLS_CHECK: "TLS_CHECK",
  SSH_CHECK: "SSH_CHECK",
} as const;