// frontend/src/api/generated/model/sshCheckResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CloudOps API
 * Infrastructure resource management, monitoring and asynchronous tasks. Timestamps use UTC ISO-8601. Error responses share ApiError; errors contains field violations.
 * OpenAPI spec version: 0.0.1-SNAPSHOT
 */
import type { SshCheckResultAuthMethod } from "./sshCheckResultAuthMethod";

export interface SshCheckResult {
  host?: string;
  port?: number;
  username?: string;
  authMethod?: SshCheckResultAuthMethod;
  serverVersion?: string;
  responseTimeMs?: number;
}